refactor: form class component
